# api/openapi.yaml
openapi: 3.1.0
info:
  title: MyShopEdge API
  version: 0.3.0
  description: "The MyShopEdge REST API. Bookkeeping and finance for UK TikTok Shop sellers.\n\nThis document\
    \ covers the seventeen endpoints that can be built without validated\nTikTok data, plus the shared\
    \ components every other endpoint will use. The\nconventions are those already set in MSE-API-001\
    \ section 4.1 and are restated here\nas schemas so they are enforced rather than described.\n\nFour\
    \ rules govern every response in this document.\n\n1. **Money is never a decimal.** Every amount is\
    \ an integer of minor units with its\n   currency beside it. A figure of 20.84 is `{\"amount_minor\"\
    : 2084, \"currency\": \"GBP\"}`.\n   Pence do not round, and a float would make them.\n\n2. **Instants\
    \ are UTC, periods are Europe/London.** A timestamp is ISO 8601 with a Z.\n   A day or a month is\
    \ a local date, because a seller's day is a London day and a VAT\n   quarter is a London quarter.\n\
    \n3. **A figure carries its own confidence.** Any response holding a money figure the\n   seller will\
    \ act on states its basis, its period and its cost coverage, and says\n   whether it is confirmed,\
    \ estimated or incomplete. A figure that cannot be trusted\n   returns null with a reason rather than\
    \ a zero.\n\n4. **Errors are problem details with a stable code.** The `code` field is the\n   contract.\
...
    \ a subscription design and is deferred.\n"
  contact:
...
servers:
...
security:
- bearerAuth: []
...
paths:
...
  /shops/{shopId}/today:
    get:
      tags:
      - Money
      summary: Hero figure, month, shop money and what needs you
      description: Traces DSH-1 and DSH-2.
      operationId: getToday
      parameters:
      - $ref: '#/components/parameters/ShopId'
      - $ref: '#/components/parameters/BasisQuery'
      responses:
        '200':
          description: Today
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodayView'
        '304':
          $ref: '#/components/responses/NotModified'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/ForbiddenShop'
...
  /shops/{shopId}/needs-you:
    get:
      tags:
      - Integrity
      summary: Ordered action list
      description: Traces DSH-10. What the seller has to do, most consequential first.
      operationId: getNeedsYou
      parameters:
      - $ref: '#/components/parameters/ShopId'
      responses:
        '200':
          description: Actions
          content:
            application/json:
              schema:
                type: object
                required:
                - as_of
                - items
                properties:
                  as_of:
                    type: string
                    format: date-time
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/NeedsYouItem'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/ForbiddenShop'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'A JWT from the identity provider, verified against its JWKS endpoint. The account

        is resolved from the `sub` claim against `accounts.auth_subject`, never from

        anything the client sends. Any `shopId` in a path is checked against that account

        before the query runs, and a shop belonging to another account returns 403 rather

        than 404, because the shop does exist and pretending otherwise helps nobody.

        '
  parameters:
    ShopId:
      name: shopId
      in: path
      required: true
      description: The MyShopEdge shop identifier, not the TikTok shop id.
      schema:
        type: string
        format: uuid
...
    BasisQuery:
      name: basis
      in: query
      description: 'Required in effect. Defaults to `sales`, and the response always states which basis

        it used, so a client that forgets to ask still knows what it received.

        '
      schema:
        type: string
        enum:
        - sales
        - cash
        default: sales
  headers:
    ETag:
      description: Entity tag. Send back as `If-None-Match` to receive 304 when unchanged.
      schema:
        type: string
...
  responses:
    NotModified:
      description: Unchanged since the supplied ETag.
    Unauthenticated:
      description: No token, an invalid token, or an expired one.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ForbiddenShop:
      description: The shop exists but does not belong to this account.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    Money:
      type: object
      description: 'An amount in integer minor units with its currency. Never a decimal, never a

        float. 20.84 pounds is `{"amount_minor": 2084, "currency": "GBP"}`.

        '
      required:
      - amount_minor
      - currency
      properties:
        amount_minor:
          type: integer
          format: int64
          examples:
          - 2084
        currency:
          type: string
          minLength: 3
          maxLength: 3
          examples:
          - GBP
    Problem:
      type: object
      description: RFC 9457 problem details, with a stable machine-readable `code`.
      required:
      - type
      - title
      - status
      - code
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
          description: Human readable. Wording may change.
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
          description: 'The contract. Clients branch on this, never on `title` or `status` alone.

            '
          enum:
          - unauthenticated
          - token_expired
          - forbidden_shop
          - not_found
          - validation_failed
          - conflict
          - idempotency_key_conflict
          - rate_limited
          - shop_needs_reconnect
          - shop_region_unsupported
          - upstream_unavailable
          - internal_error
    Confidence:
      type: string
      description: 'How far a figure can be trusted. `confirmed` means every input is settled.

        `estimated` means at least one input is a projection. `incomplete` means a

        required input is missing, and the figure it describes will be null.

        '
      enum:
      - confirmed
      - estimated
      - incomplete
    Basis:
      type: string
      description: '`sales` recognises money when the sale happens. `cash` recognises it when TikTok

        pays out. A figure without a basis is meaningless, so every period figure states it.

        '
      enum:
      - sales
      - cash
    Period:
      type: object
      required:
      - from
      - to
      - basis
      properties:
        from:
          type: string
          format: date
          description: Local date, Europe/London.
        to:
          type: string
          format: date
          description: Local date, Europe/London. Inclusive.
        basis:
          $ref: '#/components/schemas/Basis'
...
    TodayView:
      type: object
      required:
      - as_of
      - hero
      - month
      - shop_money
      - needs_you
      properties:
        as_of:
          type: string
          format: date-time
        stale:
          type: boolean
          description: 'True exactly when `freshness.status` is `stale`. Kept for clients that

            read it before `freshness` existed. A29.3.

            '
        freshness:
          type: object
          description: 'How old the figures are, computed by the service. A29.3. Under six hours

            since the last successful sync is `fresh`, six to twenty four hours inclusive is

            `getting_old`, and over twenty four hours or never synced is `stale`. The client

            renders this and never computes it.

            '
          required:
          - status
          - last_synced_at
          properties:
            status:
              type: string
              enum:
              - fresh
              - getting_old
              - stale
            last_synced_at:
              type:
              - string
              - 'null'
              format: date-time
        hero:
          type: object
          required:
          - label
          - value
          - confidence
          properties:
            label:
              type: string
              examples:
              - left_after_tiktok
            value:
              type:
              - object
              - 'null'
              allOf:
              - $ref: '#/components/schemas/Money'
            confidence:
              $ref: '#/components/schemas/Confidence'
            cost_coverage:
              type: number
              minimum: 0
              maximum: 1
        month:
          type: object
          properties:
            period:
              $ref: '#/components/schemas/Period'
            gross:
              $ref: '#/components/schemas/Money'
            net_proceeds:
              $ref: '#/components/schemas/Money'
            kept:
              type:
              - object
              - 'null'
              allOf:
              - $ref: '#/components/schemas/Money'
            kept_reason:
              type:
              - string
              - 'null'
        shop_money:
          type: object
          description: 'All time, A29.7. `generated` is net proceeds as A8 defines it. Return

            postage TikTok deducts is stated as `return_postage`, so paid out plus awaiting

            equals generated plus return postage, and paid out equals what TikTok actually

            paid.

            '
          properties:
            generated:
              $ref: '#/components/schemas/Money'
            paid_out:
              $ref: '#/components/schemas/Money'
            awaiting:
              $ref: '#/components/schemas/Money'
            return_postage:
              allOf:
              - $ref: '#/components/schemas/Money'
              description: Return postage TikTok deducted in its statements, negative, reconciled
                rather than dropped. A29.7.
            awaiting_breakdown:
              type: array
              items:
                type: object
                required:
                - status
                - amount
                - orders
                properties:
                  status:
                    type: string
                    enum:
                    - waiting_delivery
                    - waiting_return_refund
                    - delivered_awaiting_settlement
                    description: TikTok's own sub-statuses, not MyShopEdge inventions.
                  amount:
                    $ref: '#/components/schemas/Money'
                  orders:
                    type: integer
                    minimum: 0
        needs_you:
          type: array
          items:
            $ref: '#/components/schemas/NeedsYouItem'
    NeedsYouItem:
      type: object
      required:
      - type
      - count
      - severity
      properties:
        type:
          type: string
          examples:
          - returns_to_check
          - missing_costs
          - open_discrepancies
          - connection_expiring
          - unmapped_fees
        count:
          type: integer
          minimum: 0
        severity:
          type: string
          enum:
          - info
          - warning
          - critical
        label:
          type: string
        amount_at_stake:
          type:
          - object
          - 'null'
          allOf:
          - $ref: '#/components/schemas/Money'
        href:
          type:
          - string
          - 'null'
...